+++ Back/DocePecado.Persistence/Contracts/IProductPersist.cs
@@ -7,6 +7,6 @@ namespace DocePecado.Persistence.Contracts
     {
         Task<Product[]> GetAllProductsAsync(bool includeOrders = false);
         Task<Product[]> GetAllProductsByNameAsync(string name, bool includeOrders = false);
-        Task<Product> GetProductByIdAsync(int productId, bool includeOrders = false);
+        Task<Product> GetProductByIdAsync(long productId, bool includeOrders = false);
     }
 }
